requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
requests==2.32.3
chromadb==0.5.5
sentence-transformers==3.1.0
openai-whisper==20231117
pypdf==5.0.0
langchain-text-splitters==0.3.0
pydantic==2.9.2
torch==2.10.0
